END of the Date commit
NEED TO DO:

imageform ajax for project.

I think the drop down for product is not scale if we have a few thousand product,
and it makes difficult to use as well

I am trying to use ajax to do autocomplete.

more info is on http://www.gridshore.nl/blog/index.php?/archives/52-Creating-an-autocomplete-with-Spring-and-DWR.html

will finished this tomorrow