Things has been Done:
1). add a select product drop down to the form. if the person click add ingredient from the product list, the product should be preset in this ingredient form. (how do we do this? if the product hasn't been added yet) i will add this to issue.

2). Add the Add ingredient link in the productlist.  the link show only shows up when the product is the food product.

NOTE:
this current structure wont be able to add multiple ingredient to one single food products. this is something we will do when the product adding page is ready to go. at this moment, the product page cant be added, i see some exception. The conversion thingy will be added next commit.