Added ShippingAddress POJO.
the workflow as below.

once a user finished shipping, and click on Checkout.

it will bring it to the screen with Shipping Address.

once he fill out the shipping address and click on Submit,
it will send out two emails.

one is to the customer for telling them we have successfully receive his order,
one is for the owner,so he can login to the web screen and start packing the customer's order.

once the owner finished packing and the product shipped, the owner update the cart status to shipped.

in the mean time, the customer still able to track their information by logging to his/her account.