fixed the Issue r92.
it is a hack basically.

IE doesn't support the table css border-collapse="collapse" in the div pop up.

so what I did is to set the border-collapse="seperate" at loading.

when doing the show product popup. I use the javascript to override the border-collapse="collapse".

this works well on all the internet browsers.